### 2019-04-23 #1406 @whedon
[PRE REVIEW]: fmcmc: A friendly MCMC framework
**Submitting author:** @gvegayon (<a href="http://orcid.org/0000-0002-3171-0844">George Vega Yon</a>)
**Repository:** <a href="https://github.com/USCbiostats/fmcmc" target ="_blank">https://github.com/USCbiostats/fmcmc</a>
**Version:** 0.1-0
**Reviewers:** fabian-s
