### 2019-04-30 #1406 @arokem
👋 @pboesu : would you be willing to review this article for [JOSS](http://joss.theoj.org/)?
