### 2019-05-03 #1406 @arokem
👋 @fabian-s: would you be willing to review this article for [JOSS](http://joss.theoj.org)?
