### 2019-05-03 #1427 @whedon
[REVIEW]: fmcmc: A friendly MCMC framework
**Submitting author:** @gvegayon (<a href="http://orcid.org/0000-0002-3171-0844">George Vega Yon</a>)
**Repository:** <a href="https://github.com/USCbiostats/fmcmc" target ="_blank">https://github.com/USCbiostats/fmcmc</a>
**Version:** 0.2-0

- [x] **Version:** 0.2-0
