### 2019-05-10 USCbiostats/fmcmc#5 @fabian-s
LICENSE file is not standards-compliant
see checklist at openjournals/joss-reviews#1427
