### 2019-05-10 USCbiostats/fmcmc#6 @fabian-s
there is no tagged release version connected to the JOSS submission
see openjournals/joss-reviews#1427
