### 2019-05-10 USCbiostats/fmcmc#7 @fabian-s
proposal argument of kernel_new is not documented clearly
...based on the doc, I have no idea what properties the environment this operates on is expected to have.
I could make some educated guesses based on the example, but I think this needs to be made much more explicit.
