### 2019-05-10 USCbiostats/fmcmc#8 @fabian-s
DOIs of references in paper.md are missing
see checklist at openjournals/joss-reviews#1427
