### 2019-05-10 #1427 @fabian-s
LICENSE is a placeholder: USCbiostats/fmcmc#5

tagged release version is missing: USCbiostats/fmcmc#6

I don't see a clear "statement of need and audience" in the repo.

Documentation issue for proposal: USCbiostats/fmcmc#7

Another "documentation" issue I can't pin on any specific function is the more general question on how to specify any priors? MCMC only takes a "log-likelihood" as its main argument, your examples do not seem to use any. That seems very weird and impractical to me, but possibly I just misunderstand...?

DOIs are missing in references:  USCbiostats/fmcmc#8

You have a "statement of need" in your paper which I find rather hyperbolic, TBH. The ultimate flexibility you claim comes at the cost of users being required to 1) implement the (log-)likelihood/posterior themselves, and 2) figuring out a suitable scaling of the proposal (kernel) for reasonable accceptance rates, which will be challenging in almost any interesting application. In addition, it seems like a stretch to claim a "pure R" implementation if some of the heavy lifting is actually being done in C++. Could the last paragraph be phrased more accurately?

Typos: "this alows" --> allows, "convergance" --> convergence, "existing mcmc packages" --> MCMC packages
