### 2019-05-10 19474ff
Addressing #5 and dependency on Rcpp (figured out a way to simplify the reflective kernel).
