+++ src/RcppExports.cpp
@@ -1,31 +1,0 @@
-// Generated by using Rcpp::compileAttributes() -> do not edit by hand
-// Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
-
-#include <Rcpp.h>
-
-using namespace Rcpp;
-
-// normal_prop
-NumericVector normal_prop(const NumericVector& x, const NumericVector& lb, const NumericVector& ub, const NumericVector& scale, const IntegerVector& fixed);
-RcppExport SEXP _fmcmc_normal_prop(SEXP xSEXP, SEXP lbSEXP, SEXP ubSEXP, SEXP scaleSEXP, SEXP fixedSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::traits::input_parameter< const NumericVector& >::type x(xSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type lb(lbSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type ub(ubSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type scale(scaleSEXP);
-    Rcpp::traits::input_parameter< const IntegerVector& >::type fixed(fixedSEXP);
-    rcpp_result_gen = Rcpp::wrap(normal_prop(x, lb, ub, scale, fixed));
-    return rcpp_result_gen;
-END_RCPP
-}
-
-static const R_CallMethodDef CallEntries[] = {
-    {"_fmcmc_normal_prop", (DL_FUNC) &_fmcmc_normal_prop, 5},
-    {NULL, NULL, 0}
-};
-
-RcppExport void R_init_fmcmc(DllInfo *dll) {
-    R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
-    R_useDynamicSymbols(dll, FALSE);
-}
