+++ R/RcppExports.R
@@ -1,7 +1,0 @@
-# Generated by using Rcpp::compileAttributes() -> do not edit by hand
-# Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
-
-normal_prop <- function(x, lb, ub, scale, fixed) {
-    .Call('_fmcmc_normal_prop', PACKAGE = 'fmcmc', x, lb, ub, scale, fixed)
-}
-
