+++ R/imports.r
@@ -1,8 +1,6 @@
-#' @importFrom Rcpp evalCpp
 #' @importFrom coda mcmc mcmc.list
 #' @importFrom parallel makePSOCKcluster stopCluster clusterExport clusterEvalQ
 #'   clusterApply detectCores clusterSetRNGStream
-#' @useDynLib fmcmc
 #' @importFrom methods formalArgs
 #' @importFrom stats runif plogis
 NULL
