+++ README.Rmd
@@ -2,6 +2,11 @@
 output: github_document
 ---
 
+```{r include=FALSE}
+knitr::opts_chunk$set(fig.path = "man/figures/")
+```
+
+
 # fmcmc: A friendly MCMC framework <img src="man/figures/logo.png" align="right" height="140"/>
 
 [![Travis-CI Build Status](https://travis-ci.org/USCbiostats/fmcmc.svg?branch=master)](https://travis-ci.org/USCbiostats/fmcmc)
@@ -9,7 +14,7 @@ output: github_document
 [![Coverage Status](https://img.shields.io/codecov/c/github/USCbiostats/fmcmc/master.svg)](https://codecov.io/github/USCbiostats/fmcmc?branch=master)
 [![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
 [![CRAN status](https://www.r-pkg.org/badges/version/fmcmc)](https://cran.r-project.org/package=fmcmc)
-
+ [![status](http://joss.theoj.org/papers/2e86b709451443990c1c6776ebb7f756/status.svg)](http://joss.theoj.org/papers/2e86b709451443990c1c6776ebb7f756)
 
 Current features:
 
