+++ NAMESPACE
@@ -14,7 +14,6 @@ export(convergence_heildel)
 export(kernel_new)
 export(kernel_normal)
 export(kernel_reflective)
-importFrom(Rcpp,evalCpp)
 importFrom(coda,mcmc)
 importFrom(coda,mcmc.list)
 importFrom(methods,formalArgs)
@@ -27,4 +26,3 @@ importFrom(parallel,makePSOCKcluster)
 importFrom(parallel,stopCluster)
 importFrom(stats,plogis)
 importFrom(stats,runif)
-useDynLib(fmcmc)
