+++ DESCRIPTION
@@ -22,12 +22,10 @@ Suggests:
     rmarkdown,
     mcmc
 Imports:
-  Rcpp,
   parallel,
   coda,
   stats,
   methods
-LinkingTo: Rcpp
 RoxygenNote: 6.1.1
 Roxygen: list(markdown = TRUE)
 VignetteBuilder: knitr

[…]
