### 2019-05-18 af2d175
Working on sequential or joint adoption
