+++ R/mcmc.r
@@ -2,7 +2,7 @@
 #' 
 #' A flexible implementation of the Metropolis-Hastings MCMC algorithm.
 #' 
-#' @param fun A function. Returns the log-likelihood
+#' @param fun A function. Returns the log-likelihood.
 #' @param initial A numeric vector or matrix with as many rows as chains with
 #' initial values of the parameters for each chain (See details).
 #' @param nsteps Integer scalar. Length of each chain.
@@ -16,8 +16,6 @@
 #' @param ... Further arguments passed to \code{fun}.
 #' @param conv_checker A function that receives an object of class [coda::mcmc.list],
 #' and returns a logical value with `TRUE` indicating convergence.
-#' @param autostop Integer scalar. Controls the frequency with which convergence
-#' is checked (see details).
 #' 
 #' @details This function implements MCMC using the Metropolis-Hastings ratio with
 #' flexible transition kernels. Users can specify either one of the available
@@ -60,15 +58,10 @@
 #' 
 #' @section Automatic stop:
 #' 
-#' When `autostop` is greater than 0, the function will perform a convergence
-#' check every `autostop` iterations. By default, depending on the number of chains,
-#' the convergence check is done 
-#' using either the Gelman diagostic as implemented in [coda::gelman.diag]
-#' (if `nchains > 1`), or the Geweke diagnostic as implemented in [coda::geweke.diag]
-#' (if `nchains == 1`).
-#' 
-#' The user may provide a different convergence criteria by passing a different
-#' function via `conv_checker`. For more information see [convergence-checker].
+#' By default, no automatic stop is implemented. If one of the functions in 
+#' [convergence-checker] is used, then the MCMC is done by bulks as specified
+#' by the convergence checker function, and thus the algorithm will stop if,
+#' the `conv_checker` returns `TRUE`. For more information see [convergence-checker].
 #' 
 #' @return An object of class [coda::mcmc] from the \CRANpkg{coda}
 #' package. The \code{mcmc} object is a matrix with one column per parameter,
@@ -196,8 +189,6 @@
 #' @name MCMC
 NULL
 
-
-
 #' @export
 #' @rdname MCMC
 MCMC <- function(
@@ -208,10 +199,9 @@ MCMC <- function(
   nchains      = 1L,
   thin         = 1L,
   kernel       = kernel_normal(),
-  burnin       = floor(nsteps/2L),
+  burnin       = 0L,
   multicore    = FALSE,
-  conv_checker = convergence_auto(),
-  autostop     = 500L,
+  conv_checker = NULL, 
   cl           = NULL
   ) {
   
@@ -222,13 +212,12 @@ MCMC <- function(
   # Checking initial argument
   initial <- check_initial(initial, nchains)
   
-  # Checking frequency to measure batch
-  if (!is.numeric(autostop))
-    stop("The `autostop` parameter must be a number. autostop=", autostop,
+  if (multicore && nchains == 1L) 
+    stop("When `multicore = TRUE`, `nchains` should be greater than 1.",
          call. = FALSE)
-  else if (length(autostop) != 1L)
-    stop("The `autostop` parameter must be of length 1. autostop=", autostop,
-         call. = FALSE)
+  
+  if (nchains < 1L)
+    stop("`nchains` must be an integer greater than 1.", call. = FALSE)
   
     # Checkihg burnins
   if (burnin >= nsteps)
@@ -242,7 +231,7 @@ MCMC <- function(
     stop("-thin- should be >= 1.", call. = FALSE)
   
   # Filling the gap on parallel
-  if (multicore && (nchains > 1L) && !length(cl)) {
[…]
