+++ tests/testthat/test-checks.r
@@ -12,18 +12,18 @@ test_that("Error passing arguents", {
   
 })
 
-test_that("Checking errors", {
-  
-  ll <- function(p) {
-    ans <- sum(log(dnorm(y - p[1] - x*p[2], sd = p[3])))
-    if (!is.finite(ans))
-      return(-.Machine$double.xmax*1e-10)
-    ans
-  }
-  
-  expect_error(MCMC(ll, initial = 1, nsteps = 2000, autostop = "1"), "must be a number")
-  expect_error(MCMC(ll, initial = 1, nsteps = 2000, autostop = c(1,1)), "must be of length")
-})
+# test_that("Checking errors", {
+#   
+#   ll <- function(p) {
+#     ans <- sum(log(dnorm(y - p[1] - x*p[2], sd = p[3])))
+#     if (!is.finite(ans))
+#       return(-.Machine$double.xmax*1e-10)
+#     ans
+#   }
+#   
+#   expect_error(MCMC(ll, initial = 1, nsteps = 2000), "must be a number")
+#   expect_error(MCMC(ll, initial = 1, nsteps = 2000), "must be of length")
+# })
 
 # ------------------------------------------------------------------------------
 test_that("Check of initial values", {
