### 2019-05-21 24186a9
Fixing bug in MCMC when length(logratio) > 1. Changing the order of parameters and creating methods for mcmc and mcmc.list. Adding new vignette
