### 2019-05-23 1ecc702
Improving the README.Rmd file (more informative). New uniform kernel.
