+++ paper/paper.md
@@ -25,11 +25,17 @@ affiliations:
 
 # Summary
 
-Markov Chain Monte Carlo (MCMC) is used in a variety of statistical and computational venues such as: statistical inference, Markov quadrature (also known as Monte Carlo integration), stochastic optimization, among others. The **fmcmc** R [@R] package provides a flexible framework for implementing MCMC methods that use the Metropolis-Hastings algorithm [@hastings1970; @metropolis1953]. It allows a greater degree of flexibility than the widely-used metrop() function.  **fmcmc** provides the following out-of-the-box features that can be valuable for both practitioners of MCMC and educators: seamless efficient multiple-chain sampling using parallel computing, user-defined transition kernels, and automatic stop using convergence monitoring.
+Markov Chain Monte Carlo (MCMC) is used in a variety of statistical and computational venues such as: statistical inference, Markov quadrature (also known as Monte Carlo integration), stochastic optimization, among others. The **fmcmc** R [@R] package provides a flexible framework for implementing MCMC methods that use the Metropolis-Hastings algorithm [@hastings1970; @metropolis1953]. **fmcmc** provides the following out-of-the-box features that can be valuable for both practitioners of MCMC and educators:
 
-In the case of transition kernels, users can either use one of the transition kernels shipped with the package (e.g. the gaussian kernel and its bounded version, the gaussian kernel with reflective boundaries), this alows a degree of flexibility that is not possible with existing mcmc packages.
+*  Seamless efficient multiple-chain sampling using parallel computing,
 
-In addition, we allow users to automatically check for convergence during execution, and stop the MCMC run once convergance has been reached, rather than having to pre-determine a fixed number of iterations. Users can either use one of the convergence monitoring checking functions that are part of the package, for example: The Gelman and Rubin's [@Gelman1992], Geweke's [@Geweke1991], etc. Or build their own to be used within the framework.
+*  User-defined transition kernels, and
+
+*  Automatic stop using convergence monitoring.
+
+In the case of transition kernels, users can either use one of the transition kernels shipped with the package (e.g. the gaussian kernel and its bounded version, the gaussian kernel with reflective boundaries), this allows a degree of flexibility that is not possible with existing MCMC packages.
+
+The main function automatically checks convergence during execution, and stop the MCMC run once convergence has been reached, rather than having to pre-determine a fixed number of iterations. Users can either use one of the convergence monitoring checking functions that are part of the package, for example: The Gelman and Rubin's [@Gelman1992], Geweke's [@Geweke1991], etc. Or build their own to be used within the framework.
 
 While there are several other R packages that either implement MCMC algorithms or provide wrappers for implementations in other languages [see for example: @Sturtz2005; @Morey2009; @Stan2018; @Geyer2019; @Scheidegger2019], to our knowledge, the `fmcmc` package is the first one to provide a framework as flexible as the one described here and to be implemented fully within R.
 
