+++ NAMESPACE
@@ -17,6 +17,7 @@ export(convergence_heildel)
 export(kernel_new)
 export(kernel_normal)
 export(kernel_reflective)
+export(kernel_unif)
 importFrom(coda,mcmc)
 importFrom(coda,mcmc.list)
 importFrom(methods,formalArgs)

[…]
