### 2019-05-23 #1427 @gvegayon
Dear @fabian-s,

I've addressed most of your comments on https://github.com/USCbiostats/fmcmc/pull/9 (thanks for them!). Regarding the rest of the comments:

> Another "documentation" issue I can't pin on any specific function is the more general question on how to specify any priors? MCMC only takes a "log-likelihood" as its main argument, your examples do not seem to use any. That seems very weird and impractical to me, but possibly I just misunderstand...?

There is no need to have a special argument for that. Any set of particular priors can be added directly in the objective function. For example, suppose that your log-likelihood function looks like this:

```
ll <- function(p, X., y.) {

  sum(dnorm(y. - (p[1] + X.*p[2]), sd = p[3], log = TRUE))

}
```

If you want to add priors to the third parameter, say you think it is distributed Beta with parameters 2 and 8, you can do it by setting:

```
ll <- function(p, X., y.) {

  sum(dnorm(y. - (p[1] + X.*p[2]), sd = p[3], log = TRUE)) +
    dbeta(p[3], 2, 8, log = TRUE)

}
```

So there is no need to do anything else than that. The algorithm will just work.

> You have a "statement of need" in your paper which I find rather hyperbolic, TBH. The ultimate flexibility you claim comes at the cost of users being required to 1) implement the (log-)likelihood/posterior themselves, and 2) figuring out a suitable scaling of the proposal (kernel) for reasonable accceptance rates, which will be challenging in almost any interesting application

I've added a statement of need as an introduction in the README.md file in which I explicitly state who should be using this R package. This, as other MCMC packages, are for those who want to implement their own model via a log-likelihood function (for example). In particular:

1. Yes, that is exactly what you are supposed to do in this type of packages:

2. That is the idea of the personalizable transition kernel, the users can add that by themselves if they need to (also, will probably add a kernel with automatic adjustment in the future, thanks for the idea :)!)

>  In addition, it seems like a stretch to claim a "pure R" implementation if some of the heavy lifting is actually being done in C++. Could the last paragraph be phrased more accurately?

That was a very good point. We actually figured out how to rewrite a function (which was in C++) to be as efficient using only base R. Right now the package doesn't need to be compiled anymore.

Please let me know if you have any additional comments :).
