### 2019-06-24 #1427 @fabian-s
Sorry for the delay, I was travelling. Seems fine now. Mayybe add a DOI to the JSS R2WinBUGS paper as well?
