### 2019-06-24 b749c04
Adding missing DOIs (and new relevant citation) to bib (see #8)
