+++ paper/paper.bib
@@ -28,6 +28,21 @@
 }
 
 
+@article{Martin2011,
+   author = {Andrew Martin and Kevin Quinn and Jong Hee Park},
+   title = {MCMCpack: Markov Chain Monte Carlo in R},
+   journal = {Journal of Statistical Software, Articles},
+   volume = {42},
+   number = {9},
+   year = {2011},
+   keywords = {},
+   abstract = {We introduce MCMCpack, an R package that contains functions to perform Bayesian inference using posterior simulation for a number of statistical models. In addition to code that can be used to fit commonly used models, MCMCpack also contains some useful utility functions, including some additional density functions and pseudo-random number generators for statistical distributions, a general purpose Metropolis sampling algorithm, and tools for visualization.},
+   issn = {1548-7660},
+   pages = {1--21},
+   doi = {10.18637/jss.v042.i09},
+   url = {https://www.jstatsoft.org/v042/i09}
+}
+
 @Manual{Geyer2019,
     title = {mcmc: Markov Chain Monte Carlo},
     author = {Charles J. Geyer and Leif T. Johnson},
@@ -74,6 +89,22 @@ Sampler},
     url = {http://mc-stan.org/},
   }  
 
+
+@article{Stan2017,
+title = {Stan : A Probabilistic Programming Language},
+author = {Carpenter, Bob and Gelman, Andrew and Hoffman, Matthew D. and Lee, Daniel and Goodrich, Ben and Betancourt, Michael and Brubaker, Marcus and Guo, Jiqiang and Li, Peter and Riddell, Allen},
+abstractNote = {Stan is a probabilistic programming language for specifying statistical models. A Stan program imperatively defines a log probability function over parameters conditioned on specified data and constants. As of version 2.14.0, Stan provides full Bayesian inference for continuous-variable models through Markov chain Monte Carlo methods such as the No-U-Turn sampler, an adaptive form of Hamiltonian Monte Carlo sampling. Penalized maximum likelihood estimates are calculated using optimization methods such as the limited memory Broyden-Fletcher-Goldfarb-Shanno algorithm. Stan is also a platform for computing log densities and their gradients and Hessians, which can be used in alternative algorithms such as variational Bayes, expectation propagation, and marginal inference using approximate integration. To this end, Stan is set up so that the densities, gradients, and Hessians, along with intermediate quantities of the algorithm such as acceptance probabilities, are easily accessible. Stan can also be called from the command line using the cmdstan package, through R using the rstan package, and through Python using the pystan package. All three interfaces support sampling and optimization-based inference with diagnostics and posterior analysis. rstan and pystan also provide access to log probabilities, gradients, Hessians, parameter transforms, and specialized plotting.},
+doi = {10.18637/jss.v076.i01},
+journal = {Journal of Statistical Software},
+issn = {1548-7660},
+number = 1,
+volume = 76,
+place = {United States},
+year = {2017},
+month = {1}
+}
+
+
 @article{Gelman1992,
   title={Inference from iterative simulation using multiple sequences},
   author={Gelman, Andrew and Rubin, Donald B and others},
@@ -103,5 +134,6 @@ Sampler},
     number = {3},
     volume = {12},
     url = {http://www.jstatsoft.org},
+    doi = {10.18637/jss.v012.i03}
   }
 
