### 2019-06-24 #1427 @gvegayon
Hey

> Sorry for the delay, I was travelling. Seems fine now. Mayybe add a DOI to the JSS R2WinBUGS paper as well?

Done, and added another MCMC R package/paper that was relevant for the literature review. @fabian-s I will close the issues you started. Thanks!
