### 2019-06-24 #1427 @gvegayon
Also, the travis-ci build is failing in Mac due to an issue regarding LaTeX (which I cannot control). Windows and Ubuntu is building OK.
