### 2019-06-27 #1427 @arokem
@fabian-s : are all your comments addressed? I see that the "statement of need" box is still unchecked.

@gvegayon : what is the issue with the CI build? Is this something related to the JOSS paper document?
