### 2019-06-28 3236626
Addressing openjournals/joss-reviews#1427
