+++ .travis.yml
@@ -6,7 +6,7 @@ os:
   - linux
   - osx
   
-osx_image: xcode7.3
+osx_image: xcode10.2
 
 env:
  global:
