### 2019-06-28 #1427 @gvegayon
> @gvegayon : what is the issue with the CI build? Is this something related to the JOSS paper document?

No, it was something related to the version of OSX I was using. Using a more recent one fixes the issue (confirmed [here](https://travis-ci.org/USCbiostats/fmcmc/builds/551951767)).

@fabian-s I've added you to the list of authors of the R package with the role "rev" (Reviewer). LMK if that's OK with you (pulled your ORCID from one of your papers).
