### 2019-07-07 #1427 @arokem
@gvegayon : looks like all is well here. I have one small comment on the manuscript. The word "Gaussian" is derived from a name and should be capitalized (appears twice, as far as I can see).

Once you correct this, could you please create an archive of the software (e.g., using https://zenodo.org) and post the DOI on this thread?
