### 2019-07-08 79817ef
Typo on paper (see https://github.com/openjournals/joss-reviews/issues/1427#issuecomment-509029993) [skip ci]
