### 2019-07-08 2b897e6
Bump version for JOSS openjournals/joss-reviews#1427 [skip ci]
