### 2019-07-08 #1427 @gvegayon
> Done! Here is the DOI 10.5281/zenodo.3272759 Thanks!

And the new version is 0.2-0
