### 2019-07-08 #1427 @arokem
@openjournals/joss-eics : I believe this one is ready to 🚢
