### 2019-07-08 #1427 @gvegayon
> @openjournals/joss-eics : I believe this one is ready to

Just in case, I do have an issue with my last name, I have two of them! "Vega Yon" and not "Yon" only (see [this issue](https://github.com/openjournals/joss-reviews/issues/1493#issuecomment-505993965_))
