### 2019-07-08 #1427 @arokem
Oh - sorry about that! Maybe this https://github.com/USCbiostats/fmcmc/pull/10 will help?
