### 2019-07-08 #1427 @gvegayon
I don't think it works. See [here](https://github.com/openjournals/joss-reviews/issues/1493#issuecomment-506021250).
