### 2019-07-09 #1427 @arfon
> I think @arfon needs to do this manually, unfortunately.

Done.
